Solve crashes and add new features
Entire UI should hopefully properly scale to text contents. This was done so that future new languages are easy to add.

Language support has been added. All strings have been moved to resx files to allow them to be easily localized. Language will read current OS language and use that or an override can be selected from the settings menu.

Removed unused usings.

Moved all dialog boxes over to custom UI dialog boxes instead of MessageShow to control the text in the buttons.

Get the launcher filename to solve a crash when updating.

Holding down the ALT key and clicking any browse button will take the user to the root of Halo 2 documents instead the the Halo 2 map editor folder.

Add code to remove rules that aren't related to "Everyone" when doing folder permissions.

Add an option to set folder permissions manually.

If the ALT key is held when a map is packaged then the map will be cut instead of copied.

Tell the user there is no directory linked if the path is empty for models.

Make quotes work for all model commands.

Change .NET version to 4.0 from 4.6.1 to workaround an issue where it would crash if the user had 4.0 installed.